@@ -132,7 +132,7 @@ class _NWebViewState extends State<NWebView> {
         name: 'Toaster',
         onMessageReceived: (JavascriptMessage message) {
           // ignore: deprecated_member_use
-          Scaffold.of(context).showSnackBar(
+          ScaffoldMessenger.of(context).showSnackBar(
             SnackBar(content: Text(message.message)),
           );
         });
@@ -147,7 +147,7 @@ class _NWebViewState extends State<NWebView> {
               onPressed: () async {
                 final url = (await controller.data!.currentUrl())!;
                 // ignore: deprecated_member_use
-                Scaffold.of(context).showSnackBar(
+                ScaffoldMessenger.of(context).showSnackBar(
                   SnackBar(content: Text('Favorited $url')),
                 );
               },
@@ -311,7 +311,7 @@ class SampleMenu extends StatelessWidget {
   Future<void> _onListCookies(WebViewController controller, BuildContext context) async {
     final cookies = await controller.evaluateJavascript('document.cookie');
     // ignore: deprecated_member_use
-    Scaffold.of(context).showSnackBar(SnackBar(
+    ScaffoldMessenger.of(context).showSnackBar(SnackBar(
       content: Column(
         mainAxisAlignment: MainAxisAlignment.end,
         mainAxisSize: MainAxisSize.min,
@@ -327,7 +327,7 @@ class SampleMenu extends StatelessWidget {
     await controller.evaluateJavascript(
         'caches.open("test_caches_entry"); localStorage["test_localStorage"] = "dummy_entry";');
     // ignore: deprecated_member_use
-    Scaffold.of(context).showSnackBar(SnackBar(
+    ScaffoldMessenger.of(context).showSnackBar(SnackBar(
       content: Text('Added a test entry to cache.'),
     ));
   }
@@ -341,7 +341,7 @@ class SampleMenu extends StatelessWidget {
   Future<void> _onClearCache(WebViewController controller, BuildContext context) async {
     await controller.clearCache();
     // ignore: deprecated_member_use
-    Scaffold.of(context).showSnackBar(SnackBar(
+    ScaffoldMessenger.of(context).showSnackBar(SnackBar(
       content: Text("Cache cleared."),
     ));
   }
@@ -353,7 +353,7 @@ class SampleMenu extends StatelessWidget {
       message = 'There are no cookies.';
     }
     // ignore: deprecated_member_use
-    Scaffold.of(context).showSnackBar(SnackBar(
+    ScaffoldMessenger.of(context).showSnackBar(SnackBar(
       content: Text(message),
     ));
   }
@@ -402,7 +402,7 @@ class NavigationControls extends StatelessWidget {
                   await controller.goBack();
                 } else {
                   // ignore: deprecated_member_use
-                  Scaffold.of(context).showSnackBar(
+                  ScaffoldMessenger.of(context).showSnackBar(
                     SnackBar(content: Text("No back history item")),
                   );
                   return;
@@ -421,7 +421,7 @@ class NavigationControls extends StatelessWidget {
                     await controller.goForward();
                   } else {
                     // ignore: deprecated_member_use
-                    Scaffold.of(context).showSnackBar(
+                    ScaffoldMessenger.of(context).showSnackBar(
                       SnackBar(
                           content: Text("No forward history item")),
                     );
